refactor(attribution): extract oversized CLI test module
Keep attribution tests in their single owning sibling module once their
inline block exceeds the repository limit. Test behavior is unchanged.
